Ajusta pacotes da api Aero Data Box na consulta de passagens aereas e aeroportos
