+++ api/wanderlust/src/main/java/com/wanderlust/wanderlust/external/airportAndFlight/aeroDataBox/model/aiportRoutesDailyFlight/AirportOperators.java
@@ -1,0 +1,8 @@
+package com.wanderlust.wanderlust.external.airportAndFlight.aeroDataBox.model.aiportRoutesDailyFlight;
+
+import lombok.Data;
+
+@Data
+public class AirportOperators {
+    private String name;
+}
