--- api/wanderlust/src/main/java/com/wanderlust/wanderlust/external/airportAndFlight/model/aiportRoutesDailyFlight/AirportOperators.java
@@ -1,8 +1,0 @@
-package com.wanderlust.wanderlust.external.airportAndFlight.model.aiportRoutesDailyFlight;
-
-import lombok.Data;
-
-@Data
-public class AirportOperators {
-    private String name;
-}
