--- api/wanderlust/src/main/java/com/wanderlust/wanderlust/external/airportAndFlight/model/airportSearchResult/AirportSearchResult.java
@@ -1,14 +1,0 @@
-package com.wanderlust.wanderlust.external.airportAndFlight.model.airportSearchResult;
-
-import com.wanderlust.wanderlust.external.airportAndFlight.model.aiportInfo.Airport;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class AirportSearchResult {
-
-    private String searchBy;
-    private List<Airport> items;
-
-}
